+++ NEWS.md
@@ -32,6 +32,9 @@
 * `find_terms()` and `find_transformation()` now better cope with inverse
   transformations of the response value, such as `1/y`.
 
+* `get_transformation()` now returns more transformations for power-transformed
+  response variables.
+
 ## Bug fixes
 
 * `null_model()` now correctly handles zero-inflated models from package